Organizar interfaces e composables, alterar mensagens para ingles
