--- .env
@@ -1,1 +1,0 @@
-GRAPHQL_URI=https://rickandmortyapi.com/graphql
