+++ src/boot/i18n.ts
@@ -4,12 +4,11 @@ import { createI18n } from 'vue-i18n';
 import messages from 'src/i18n';
 
 const i18n = createI18n({
-  locale: 'pt-BR',
+  locale: 'en-US',
   messages,
 });
 
 export default boot(({ app }) => {
-  // Set i18n instance on app
   app.use(i18n);
 });
