+++ src/components/Loader.vue
@@ -3,3 +3,11 @@
     <q-spinner-dots color="primary" size="40px" />
   </div>
 </template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+
+export default defineComponent({
+  name: 'Loader',
+});
+</script>
