--- src/components/interfaces.ts
@@ -1,9 +1,0 @@
-/* eslint-disable no-unused-vars */
-export interface Todo {
-  id: number;
-  content: string;
-}
-
-export interface Meta {
-  totalCount: number;
-}
