+++ src/composables/utils/model.ts
@@ -1,7 +1,7 @@
 import { computed } from 'vue';
-import { Emit } from './interfaces';
+import { ModelEmit } from '../../interfaces';
 
-export function useModel(props: Readonly<Record<string, string | undefined>>, emit: Emit) {
+export function useModel(props: Readonly<Record<string, string | undefined>>, emit: ModelEmit) {
   return computed({
     get: () => props.modelValue,
     set: (value) => emit('update:modelValue', value),
