+++ src/composables/utils/routeId.ts
