+++ src/i18n/index.ts
@@ -1,5 +1,5 @@
-import ptBR from './pt-BR';
+import enUS from './en-US';
 
 export default {
-  'pt-BR': ptBR,
+  'en-US': enUS,
 };
