+++ src/index.template.html
@@ -16,7 +16,6 @@
     <link rel="icon" type="image/ico" href="favicon.ico">
   </head>
   <body>
-    <!-- DO NOT touch the following DIV -->
     <div id="q-app"></div>
   </body>
 </html>
