+++ src/interfaces/index.ts
@@ -1,0 +1,5 @@
+export * from './character';
+export * from './episode';
+export * from './infiniteScroll';
+export * from './location';
+export * from './model';
