+++ src/interfaces/location.ts
@@ -1,0 +1,3 @@
+export interface Location {
+  name: string;
+}
