+++ src/interfaces/model.ts
@@ -1,0 +1,5 @@
+/* eslint-disable no-unused-vars */
+
+export interface ModelEmit {
+  (event: string, payload: string | undefined): void;
+}
